Fix State.Init
When starting the node with an existing database, we were looking
for the state root in the last transaction's receipt. This was
making us miss root when there were no transactions and
CreateAccounts had updated the root. So Commit now comutes the root
using a special function, saves the root in a special key, and
Init looks for that value.