Always initialize form empty state, not last root
We use to keep track of the latest state hash (root in evm parlance)
in the chaindata db, and initialize the state from that point if a
database already existed. But Babble replays all the past blocks when
it is restarted with an exisiting Babble database, so we were
re-applying transactions to an already up-to-date state, messing up the
system. Basically, the problem is that there are two databases: one for
Babble, and one for the chaindata. Which one do we use to bootstrap a
node? We chose the Babble DB, replaying transactions on an empty state.