Update docker deploy scripts with 'evml keys' cmd
Now that Babble uses the same ECDSA curve as Ethereum (secp256k1),
we can use the same key across both. The deploy scripts are updated
to generate Ethereum keys and to reuse them as the validator key
for Babble, using the new evml keys 'generate' and 'inspect'
commands. The corresponding accounts are pre-funded in the genesis
block. Hence Babble validators have funds in the EVM-Lite ledger,
which they control with the same private key they use in Babble
consensus.