WIP demopoa
The demopoa folder contains some handmade configuration for 4 nodes.

The babble genesis peer-set contains only node1. The idea is to then add nodes
2-4 via the POA/DynamicMembership method. The POA contract is initialized with
the keys of node1 and node2, so node2 doesnt need to go through the voting
nomination/voting process, but nodes 3 and 4 do.

I had to manually change the eth/genesis.json file to set 'authorizing=true' on
the POA smart-contract account.

Also modified the babble inmem_proxy to use the checkAuthorized method, rather
than checkAuthorizedPublicKey because the latter would actually require a 64
bytes parameter, not 32. So we convert the public key into an address, and pass
it to the checkAuthorized method of the SmartContract.

TODO:

 * Update the inmem_proxy_test
 * Improve demopoa