Delete a lot of things. This is a major change.
We decided to completely refactor evm-lite. It is not an executable
anymore, just a library. We leave it to the users of this library to
implement the consensus interface and create an executable around it.

So we removed the /cmd /e2e and /deploy folders. At Mosaic, we will
implement these separately, in the monetd repo or in a special
dev-ops repo.

We also removed the consensus/babble and consensus/raft folders.
These will also be defined in separate repos.

We dropped the functionnality of managing keys in the server. This
is was a security issue, and we would rather force users to sign
transactions locally.

The Readme has been partially amended for these changes, but it can
be improved.

The docs/ folder was deleted.